chore: fix circular dependency issues across packages
