+++ packages/advisor-components/src/RuleDetails/RuleDetailsMessages.ts
@@ -1,0 +1,16 @@
+export type Message = React.ReactNode;
+export type RuleDetailsMessages = {
+  systemReboot: Message;
+  knowledgebaseArticle: Message;
+  totalRisk: Message;
+  rulesDetailsTotalRiskBody: Message;
+  likelihoodLevel: Message;
+  likelihoodDescription: Message;
+  impactLevel: Message;
+  impactDescription: Message;
+  riskOfChange: Message;
+  riskOfChangeText: Message;
+  riskOfChangeLabel: Message;
+  ruleHelpful: Message;
+  feedbackThankYou: Message;
+};
