+++ packages/components/project.json
@@ -47,6 +47,12 @@
         "outputPath": "dist/@redhat-cloud-services/frontend-components"
       }
     },
+    "check-circular-imports": {
+      "executor": "@redhat-cloud-services/frontend-components-executors:check-circular-imports",
+      "options": {
+        "sourceDir": "packages/components/src"
+      }
+    },
     "build": {
       "executor": "nx:noop",
       "dependsOn": [
