+++ packages/components/src/ConditionalFilter/index.ts
@@ -3,7 +3,8 @@ export { default as ConditionalFilter } from './ConditionalFilter';
 export type { ConditionalFilterProps, ConditionalFilterItem, TextInputProps } from './ConditionalFilter';
 export type { FilterValues } from './ConditionalFilter';
 export { default as GroupFilter } from './GroupFilter';
-export type { GroupFilterProps, Group } from './GroupFilter';
+export type { GroupFilterProps } from './GroupFilter';
+export type { Group } from './groupFilterConstants';
 export { default as CheckboxFilter } from './CheckboxFilter';
 export type { CheckboxFilterProps } from './CheckboxFilter';
 export { default as RadioFilter } from './RadioFilter';
