+++ packages/components/src/FilterHooks/tagFilterHook.tsx
@@ -2,7 +2,7 @@ import React, { SetStateAction, useEffect, useState } from 'react';
 import { constructGroups, mapGroups } from './constants';
 import { Spinner } from '@patternfly/react-core';
 import './tagFilterHook.scss';
-import { AllTag, GroupItem, GroupValue } from './constants';
+import { AllTag, FilterHookGroupItem, GroupValue } from './constants';
 
 export const tagsFilterState = { tagsFilter: {} };
 export const TAGS_FILTER = 'TAGS_FILTER';
@@ -70,7 +70,7 @@ export const useTagsFilter: UseTagsFilter = (
       _e: any,
       newSelection: { [key: string]: GroupValue },
       group: Record<string, unknown>,
-      item: GroupItem,
+      item: FilterHookGroupItem,
       groupKey: string,
       itemKey: string
     ) => {
