+++ packages/translations/src/index.js
@@ -1,4 +1,4 @@
 export { default as IntlProvider } from './Provider';
 export { default as defaultMessages } from './messages';
-export const LOCALSTORAGE_KEY = 'rhcs-language';
+export * from './utils';
 export { default as intlHelper } from './intlHelper';
