+++ packages/translations/src/utils.js
@@ -1,0 +1,1 @@
+export const LOCALSTORAGE_KEY = 'rhcs-language';
